pyproject.toml
[tool.poetry]
name = "chat"
version = "0.1.0"
description = ""
authors = ["Michael Yoffe <myoffe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
requests = "^2.28.2"
pymongo = "^4.3.3"
aiohttp = "^3.8.4"
eventlet = "^0.33.3"
python-socketio = "^5.8.0"
aioconsole = "^0.6.1"
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
